third_party/blink/web_tests/virtual/netinfo-constant-type/netinfo/unregister-during-event-expected.txt

Tests that it's okay to unregister the event listener while handling the event.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL connection.type should be ethernet. Was unknown.
FAIL connection.downlinkMax should be 2. Was Infinity.
PASS connection.effectiveType is initialEffectiveType
PASS connection.rtt is initialRtt
PASS connection.downlink is initialDownlink
FAIL connection.type should be bluetooth. Was unknown.
FAIL connection.downlinkMax should be 1. Was Infinity.
PASS connection.effectiveType is initialEffectiveType
PASS connection.rtt is initialRtt
PASS connection.downlink is initialDownlink
PASS successfullyParsed is true

TEST COMPLETE
